Updated db connectivity and edited adapter, controller and repo.
