+++ pom.xml
@@ -33,7 +33,7 @@
 		<dependency>
 			<groupId>org.flywaydb</groupId>
 			<artifactId>flyway-core</artifactId>
-			<version>9.20.0</version>
+			<version>9.16.3</version>
 		</dependency>
 		<dependency>
 			<groupId>org.projectlombok</groupId>
