+++ src/main/java/com/mastercode/fitmaster/model/Trainer.java
@@ -1,6 +1,5 @@
-package com.mastercode.fitmaster.trainer.model.entity;
+package com.mastercode.fitmaster.model;
 
-import com.mastercode.fitmaster.model.Gender;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -14,7 +13,7 @@ import java.time.LocalDate;
 @AllArgsConstructor
 @Entity
 @Table(name = "trainers")
-public class TrainerEntity {
+public class Trainer {
     @Id
     @GeneratedValue
     @Column(name = "id")
