+++ src/main/resources/application.properties
@@ -9,3 +9,8 @@ flyway.enabled=true
 flyway.url=jdbc:postgresql://localhost:9001/fitmaster
 flyway.user=admin
 flyway.password=admin
+spring.flyway.baseline-on-migrate=true
+spring.flyway.locations=classpath:database/migrations
+
+# Logger configuration
+#logging.level.root=debug
